--- src/main/java/site/gachontable/gachontablebe/domain/auth/domain/AuthDetails.java
@@ -1,55 +1,0 @@
-package site.gachontable.gachontablebe.domain.auth.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import site.gachontable.gachontablebe.domain.shared.Role;
-
-import java.util.Collection;
-import java.util.Collections;
-import java.util.UUID;
-
-@Getter
-@AllArgsConstructor
-public class AuthDetails implements UserDetails {
-    private final UUID uuid;
-    private final String tel;
-    private final Role role;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.singletonList(new SimpleGrantedAuthority(role.getRole()));
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return tel;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
