--- src/main/java/site/gachontable/gachontablebe/domain/auth/usecase/AdminAuthDetailsService.java
@@ -1,25 +1,0 @@
-package site.gachontable.gachontablebe.domain.auth.usecase;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import site.gachontable.gachontablebe.domain.admin.domain.Admin;
-import site.gachontable.gachontablebe.domain.admin.domain.repository.AdminRepository;
-import site.gachontable.gachontablebe.domain.admin.exception.AdminNotFoundException;
-import site.gachontable.gachontablebe.domain.auth.domain.AuthDetails;
-import site.gachontable.gachontablebe.domain.shared.Role;
-
-@Service
-@RequiredArgsConstructor
-public class AdminAuthDetailsService implements UserDetailsService {
-
-    private final AdminRepository adminRepository;
-
-    @Override
-    public AuthDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Admin admin = adminRepository.findByUsername(username)
-                .orElseThrow(AdminNotFoundException::new);
-        return new AuthDetails(admin.getAdminId(), admin.getUsername(), Role.ROLE_ADMIN);
-    }
-}
