--- src/main/java/site/gachontable/gachontablebe/domain/auth/usecase/UserAuthDetailsService.java
@@ -1,25 +1,0 @@
-package site.gachontable.gachontablebe.domain.auth.usecase;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import site.gachontable.gachontablebe.domain.auth.domain.AuthDetails;
-import site.gachontable.gachontablebe.domain.shared.Role;
-import site.gachontable.gachontablebe.domain.user.domain.User;
-import site.gachontable.gachontablebe.domain.user.domain.repository.UserRepository;
-import site.gachontable.gachontablebe.domain.user.exception.UserNotFoundException;
-
-@Service
-@RequiredArgsConstructor
-public class UserAuthDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public AuthDetails loadUserByUsername(String tel) throws UsernameNotFoundException {
-        User user = userRepository.findByUserTel(tel)
-                .orElseThrow(UserNotFoundException::new);
-        return new AuthDetails(user.getUserId(), user.getUserTel(), Role.ROLE_USER);
-    }
-}
